@@ -1,13 +1,13 @@
 {
   "name": "angular-boilerplate",
   "version": "0.0.0",
-  "sideEffects": false,
   "private": true,
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:prod": "cross-env NODE_ENV=production ng build --prod --build-optimizer --aot --stats-json",
+    "build:stats": "ng build --stats-json",
+    "watch": "ng build --watch --configuration development",
     "test:unit": "ng test",
     "test:e2e": "ng e2e",
     "test:coverage": "ng test --coverage",
@@ -15,29 +15,28 @@
     "analyze": "webpack-bundle-analyzer dist/angular-boilerplate/stats.json"
   },
   "dependencies": {
-    "@angular/animations": "11.2.13",
-    "@angular/common": "11.2.13",
-    "@angular/compiler": "11.2.13",
-    "@angular/core": "11.2.13",
-    "@angular/forms": "11.2.13",
-    "@angular/platform-browser": "11.2.13",
-    "@angular/platform-browser-dynamic": "11.2.13",
-    "@angular/router": "11.2.13",
-    "@angular/service-worker": "11.2.13",
+    "@angular/animations": "~12.0.1",
+    "@angular/common": "~12.0.1",
+    "@angular/compiler": "~12.0.1",
+    "@angular/core": "~12.0.1",
+    "@angular/forms": "~12.0.1",
+    "@angular/platform-browser": "~12.0.1",
+    "@angular/platform-browser-dynamic": "~12.0.1",
+    "@angular/router": "~12.0.1",
     "rxjs": "6.6.7",
     "tslib": "2.2.0",
-    "zone.js": "0.11.4"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-builders/jest": "12.0.0",
-    "@angular-devkit/build-angular": "0.1102.13",
+    "@angular-devkit/build-angular": "~12.0.1",
     "@angular-eslint/builder": "12.0.0",
     "@angular-eslint/eslint-plugin": "12.0.0",
     "@angular-eslint/eslint-plugin-template": "12.0.0",
     "@angular-eslint/schematics": "12.0.0",
     "@angular-eslint/template-parser": "12.0.0",
-    "@angular/cli": "11.2.13",
-    "@angular/compiler-cli": "11.2.13",
+    "@angular/cli": "~12.0.1",
+    "@angular/compiler-cli": "~12.0.1",
     "@briebug/cypress-schematic": "5.0.0",
     "@commitlint/cli": "12.1.4",
     "@commitlint/config-conventional": "12.1.4",
@@ -48,7 +47,6 @@
     "@typescript-eslint/eslint-plugin": "4.24.0",
     "@typescript-eslint/parser": "4.24.0",
     "autoprefixer": "10.2.5",
-    "cross-env": "7.0.3",
     "cypress": "7.3.0",
     "eslint": "7.27.0",
     "eslint-config-prettier": "8.3.0",
@@ -67,7 +65,7 @@
     "tailwindcss": "2.1.2",
     "ts-loader": "9.2.2",
     "ts-node": "10.0.0",
-    "typescript": "4.1.5",
+    "typescript": "4.2.4",
     "webpack-bundle-analyzer": "4.4.2"
   }
 }
