@@ -1,19 +1,11 @@
 import { CommonModule } from '@angular/common';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { ServiceWorkerModule } from '@angular/service-worker';
-import { environment } from '@environments/environment';
 import { JwtInterceptor, ServerErrorInterceptor } from './interceptors';
 
 @NgModule({
   declarations: [],
-  imports: [
-    CommonModule,
-    HttpClientModule,
-    ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production,
-    }),
-  ],
+  imports: [CommonModule, HttpClientModule],
   providers: [
     {
       provide: HTTP_INTERCEPTORS,
